osqp_sys

Function osqp_setup

pub unsafe extern "C" fn osqp_setup(
    solverp: *mut *mut OSQPSolver,
    P: *const OSQPCscMatrix,
    q: *const osqp_float,
    A: *const OSQPCscMatrix,
    l: *const osqp_float,
    u: *const osqp_float,
    m: osqp_int,
    n: osqp_int,
    settings: *const OSQPSettings,
) -> osqp_int

Initialize OSQP solver allocating memory.

It performs:

data and settings validation

problem data scaling

setup linear system solver:

direct solver: KKT matrix factorization is performed here

indirect solver: reduced KKT matrix preconditioning is performed here

NB: This is the only function that allocates dynamic memory and is not used during code generation

@param solverp Solver pointer @param P Problem data (upper triangular part of quadratic cost term, csc format) @param q Problem data (linear cost term) @param A Problem data (constraint matrix, csc format) @param l Problem data (constraint lower bound) @param u Problem data (constraint upper bound) @param m Problem data (number of constraints) @param n Problem data (number of variables) @param settings Solver settings @return Exitflag for errors (0 if no errors)