osqp_sys

Function osqp_update_data_mat

pub unsafe extern "C" fn osqp_update_data_mat(
    solver: *mut OSQPSolver,
    Px_new: *const osqp_float,
    Px_new_idx: *const osqp_int,
    P_new_n: osqp_int,
    Ax_new: *const osqp_float,
    Ax_new_idx: *const osqp_int,
    A_new_n: osqp_int,
) -> osqp_int

Update elements of matrices P (upper triangular) and A by preserving their sparsity structures.

If Px_new_idx (Ax_new_idx) is OSQP_NULL, Px_new (Ax_new) is assumed to be as long as P->x (A->x) and the whole P->x (A->x) is replaced.

@param solver Solver @param Px_new Vector of new elements in P->x (upper triangular), NULL if none @param Px_new_idx Index mapping new elements to positions in P->x @param P_new_n Number of new elements to be changed @param Ax_new Vector of new elements in A->x, NULL if none @param Ax_new_idx Index mapping new elements to positions in A->x @param A_new_n Number of new elements to be changed @return output flag: 0: OK 1: P_new_n > nnzP 2: A_new_n > nnzA <0: error in the update