Struct osqp_sys::OSQPData

#[repr(C)]

pub struct OSQPData {
    pub n: osqp_int,
    pub m: osqp_int,
    pub P: *mut csc,
    pub A: *mut csc,
    pub q: *mut osqp_float,
    pub l: *mut osqp_float,
    pub u: *mut osqp_float,
}

Data structure

Fields
n: osqp_int

< number of variables n

m: osqp_int

< number of constraints m

P: *mut csc

< quadratic part of the cost P in csc format (size n x n). It

A: *mut csc

< linear constraints matrix A in csc format (size m x n)

q: *mut osqp_float

< dense array for linear part of cost function (size n)

l: *mut osqp_float

< dense array for lower bound (size m)

u: *mut osqp_float

< dense array for upper bound (size m)