Struct osqp_sys::OSQPWorkspace

#[repr(C)]pub struct OSQPWorkspace {
    pub data: *mut OSQPData,
    pub linsys_solver: *mut LinSysSolver,
    pub pol: *mut OSQPPolish,
    pub rho_vec: *mut osqp_float,
    pub rho_inv_vec: *mut osqp_float,
    pub constr_type: *mut osqp_int,
    pub x: *mut osqp_float,
    pub y: *mut osqp_float,
    pub z: *mut osqp_float,
    pub xz_tilde: *mut osqp_float,
    pub x_prev: *mut osqp_float,
    pub z_prev: *mut osqp_float,
    pub Ax: *mut osqp_float,
    pub Px: *mut osqp_float,
    pub Aty: *mut osqp_float,
    pub delta_y: *mut osqp_float,
    pub Atdelta_y: *mut osqp_float,
    pub delta_x: *mut osqp_float,
    pub Pdelta_x: *mut osqp_float,
    pub Adelta_x: *mut osqp_float,
    pub D_temp: *mut osqp_float,
    pub D_temp_A: *mut osqp_float,
    pub E_temp: *mut osqp_float,
    pub settings: *mut OSQPSettings,
    pub scaling: *mut OSQPScaling,
    pub solution: *mut OSQPSolution,
    pub info: *mut OSQPInfo,
    pub timer: *mut OSQPTimer,
    pub first_run: osqp_int,
    pub clear_update_time: osqp_int,
    pub rho_update_from_solve: osqp_int,
    pub summary_printed: osqp_int,
}

OSQP Workspace

Fields
data: *mut OSQPData

Problem data to work on (possibly scaled)

linsys_solver: *mut LinSysSolver

Linear System solver structure

pol: *mut OSQPPolish

Polish structure

rho_vec: *mut osqp_float

< vector of rho values

rho_inv_vec: *mut osqp_float

< vector of inv rho values

constr_type: *mut osqp_int

< Type of constraints: loose (-1), equality (1), inequality (0)

x: *mut osqp_float

< Iterate x

y: *mut osqp_float

< Iterate y

z: *mut osqp_float

< Iterate z

xz_tilde: *mut osqp_float

< Iterate xz_tilde

x_prev: *mut osqp_float

< Previous x

z_prev: *mut osqp_float

< Previous z

Ax: *mut osqp_float

< scaled A * x

Px: *mut osqp_float

< scaled P * x

Aty: *mut osqp_float

< scaled A' * y

delta_y: *mut osqp_float

< difference between consecutive dual iterates

Atdelta_y: *mut osqp_float

< A' * delta_y

delta_x: *mut osqp_float

< difference between consecutive primal iterates

Pdelta_x: *mut osqp_float

< P * delta_x

Adelta_x: *mut osqp_float

< A * delta_x

D_temp: *mut osqp_float

< temporary primal variable scaling vectors

D_temp_A: *mut osqp_float

< temporary primal variable scaling vectors storing norms of A columns

E_temp: *mut osqp_float

< temporary constraints scaling vectors storing norms of A' columns

settings: *mut OSQPSettings

< problem settings

scaling: *mut OSQPScaling

< scaling vectors

solution: *mut OSQPSolution

< problem solution

info: *mut OSQPInfo

< solver information

timer: *mut OSQPTimer

< timer object

first_run: osqp_int

flag indicating whether the solve function has been run before

clear_update_time: osqp_int

flag indicating whether the update_time should be cleared

rho_update_from_solve: osqp_int

flag indicating that osqp_update_rho is called from osqp_solve function

summary_printed: osqp_int

< Has last summary been printed? (true/false)