Crate osqp_sys

Structs

OSQPCscMatrix

Matrix in compressed-column form. The structure is used internally to store matrices in the triplet form as well, but the API requires that the matrices are in the CSC format.

OSQPInfo

Information about the solution process.

OSQPSettings

User settings

OSQPSolution

Structure to hold the computed solution (if any), and any certificates of infeasibility (if any) found by the solver.

OSQPSolver

Main OSQP solver structure that holds all information.

OSQPWorkspace_

Enums

OSQPTimer

Constants

OSQP_ALGEBRA_LOAD_ERROR

OSQP_CODEGEN_DEFINES_ERROR

OSQP_DATA_NOT_INITIALIZED

OSQP_DATA_VALIDATION_ERROR

OSQP_DIAGONAL_PRECONDITIONER

OSQP_DIRECT_SOLVER

OSQP_DUAL_INFEASIBLE

OSQP_DUAL_INFEASIBLE_INACCURATE

OSQP_FOPEN_ERROR

OSQP_FUNC_NOT_IMPLEMENTED

< Function not implemented in this library

OSQP_INDIRECT_SOLVER

OSQP_LAST_ERROR_PLACE

OSQP_LINSYS_SOLVER_INIT_ERROR

OSQP_MAX_ITER_REACHED

OSQP_MEM_ALLOC_ERROR

OSQP_NONCVX_ERROR

OSQP_NON_CVX

OSQP_NO_ERROR

OSQP_NO_PRECONDITIONER

OSQP_PRIMAL_INFEASIBLE

OSQP_PRIMAL_INFEASIBLE_INACCURATE

OSQP_SETTINGS_VALIDATION_ERROR

OSQP_SIGINT

OSQP_SOLVED

OSQP_SOLVED_INACCURATE

OSQP_TIME_LIMIT_REACHED

OSQP_UNKNOWN_SOLVER

OSQP_UNSOLVED

OSQP_WORKSPACE_NOT_INIT_ERROR

Functions

OSQPCscMatrix_free⚠

Free a CSC matrix object allocated by @ref OSQPCscMatrix_new.

OSQPCscMatrix_new⚠

Allocates a new Compressed-Column-Sparse (CSC) matrix from existing arrays.

OSQPSettings_free⚠

Free an OSQPSettings object.

OSQPSettings_new⚠

Allocate a new OSQPSettings object with the default settings.

osqp_cleanup⚠

Cleanup workspace by deallocating memory

osqp_set_default_settings⚠

Get the default settings from the osqp_api_constants.h file.

osqp_setup⚠

Initialize OSQP solver allocating memory.

osqp_solve⚠

Solve quadratic program

osqp_update_data_mat⚠

Update elements of matrices P (upper triangular) and A by preserving their sparsity structures.

osqp_update_data_vec⚠

Update problem data vectors @param solver Solver @param q_new New linear cost, NULL if none @param l_new New lower bound, NULL if none @param u_new New upper bound, NULL if none @return Exitflag for errors (0 if no errors)

osqp_update_rho⚠

Update the ADMM parameter rho.

osqp_update_settings⚠

Update problem settings.

osqp_version⚠

Return OSQP version

osqp_warm_start⚠

Warm start primal and dual variables @param solver Solver @param x Primal variable, NULL if none @param y Dual variable, NULL if none @return Exitflag for errors (0 if no errors)

Type Aliases

OSQPWorkspace

osqp_error_type

Solver Errors *

osqp_float

osqp_int

osqp_linsys_solver_type

Linear System Solvers *

osqp_precond_type

Preconditioners for CG method *

osqp_status_type

Solver Status *