Struct osqp_sys::OSQPSettings

#[repr(C)]pub struct OSQPSettings {
    pub rho: osqp_float,
    pub sigma: osqp_float,
    pub scaling: osqp_int,
    pub adaptive_rho: osqp_int,
    pub adaptive_rho_interval: osqp_int,
    pub adaptive_rho_tolerance: osqp_float,
    pub adaptive_rho_fraction: osqp_float,
    pub max_iter: osqp_int,
    pub eps_abs: osqp_float,
    pub eps_rel: osqp_float,
    pub eps_prim_inf: osqp_float,
    pub eps_dual_inf: osqp_float,
    pub alpha: osqp_float,
    pub linsys_solver: linsys_solver_type,
    pub delta: osqp_float,
    pub polish: osqp_int,
    pub polish_refine_iter: osqp_int,
    pub verbose: osqp_int,
    pub scaled_termination: osqp_int,
    pub check_termination: osqp_int,
    pub warm_start: osqp_int,
    pub time_limit: osqp_float,
}

Settings struct

Fields
rho: osqp_float

< ADMM step rho

sigma: osqp_float

< ADMM step sigma

scaling: osqp_int

< heuristic data scaling iterations; if 0, then disabled.

adaptive_rho: osqp_int

< boolean, is rho step size adaptive?

adaptive_rho_interval: osqp_int

< number of iterations between rho adaptations; if 0, then it is automatic

adaptive_rho_tolerance: osqp_float

< tolerance X for adapting rho. The new rho has to be X times larger or 1/X times smaller than the current one to trigger a new factorization.

adaptive_rho_fraction: osqp_float

< interval for adapting rho (fraction of the setup time)

max_iter: osqp_int

< maximum number of iterations

eps_abs: osqp_float

< absolute convergence tolerance

eps_rel: osqp_float

< relative convergence tolerance

eps_prim_inf: osqp_float

< primal infeasibility tolerance

eps_dual_inf: osqp_float

< dual infeasibility tolerance

alpha: osqp_float

< relaxation parameter

linsys_solver: linsys_solver_type

< linear system solver to use

delta: osqp_float

< regularization parameter for polishing

polish: osqp_int

< boolean, polish ADMM solution

polish_refine_iter: osqp_int

< number of iterative refinement steps in polishing

verbose: osqp_int

< boolean, write out progress

scaled_termination: osqp_int

< boolean, use scaled termination criteria

check_termination: osqp_int

< integer, check termination interval; if 0, then termination checking is disabled

warm_start: osqp_int

< boolean, warm start

time_limit: osqp_float

< maximum number of seconds allowed to solve the problem; if 0, then disabled