osqp_sys

Struct OSQPSettings

#[repr(C)]
pub struct OSQPSettings {

    pub device: osqp_int,
    pub linsys_solver: osqp_linsys_solver_type,
    pub allocate_solution: osqp_int,
    pub verbose: osqp_int,
    pub profiler_level: osqp_int,
    pub warm_starting: osqp_int,
    pub scaling: osqp_int,
    pub polishing: osqp_int,
    pub rho: osqp_float,
    pub rho_is_vec: osqp_int,
    pub sigma: osqp_float,
    pub alpha: osqp_float,
    pub cg_max_iter: osqp_int,
    pub cg_tol_reduction: osqp_int,
    pub cg_tol_fraction: osqp_float,
    pub cg_precond: osqp_precond_type,
    pub adaptive_rho: osqp_int,
    pub adaptive_rho_interval: osqp_int,
    pub adaptive_rho_fraction: osqp_float,
    pub adaptive_rho_tolerance: osqp_float,
    pub max_iter: osqp_int,
    pub eps_abs: osqp_float,
    pub eps_rel: osqp_float,
    pub eps_prim_inf: osqp_float,
    pub eps_dual_inf: osqp_float,
    pub scaled_termination: osqp_int,
    pub check_termination: osqp_int,
    pub check_dualgap: osqp_int,
    pub time_limit: osqp_float,
    pub delta: osqp_float,
    pub polish_refine_iter: osqp_int,

}

User settings

Fields
device: osqp_int

< device identifier; currently used for CUDA devices

linsys_solver: osqp_linsys_solver_type

< linear system solver to use

allocate_solution: osqp_int

< boolean; allocate solution in OSQPSolver during osqp_setup

verbose: osqp_int

< boolean; write out progress

profiler_level: osqp_int

< integer; level of detail for profiler annotations

warm_starting: osqp_int

< boolean; warm start

scaling: osqp_int

< data scaling iterations; if 0, then disabled

polishing: osqp_int

< boolean; polish ADMM solution

rho: osqp_float

< ADMM penalty parameter

rho_is_vec: osqp_int

< boolean; is rho scalar or vector?

sigma: osqp_float

< ADMM penalty parameter

alpha: osqp_float

< ADMM relaxation parameter

cg_max_iter: osqp_int

< maximum number of CG iterations per solve

cg_tol_reduction: osqp_int

< number of consecutive zero CG iterations before the tolerance gets halved

cg_tol_fraction: osqp_float

< CG tolerance (fraction of ADMM residuals)

cg_precond: osqp_precond_type

< Preconditioner to use in the CG method

adaptive_rho: osqp_int

rho stepsize adaption method

adaptive_rho_interval: osqp_int

Interval between rho adaptations

When adaptive_rho == OSQP_ADAPTIVE_RHO_UPDATE_ITERATIONS, this is the number of iterations between rho updates.

Not used when adaptive_rho is any other value.

adaptive_rho_fraction: osqp_float

Adaptation parameter controlling when non-fixed rho adaptations occur.

When adaptive_rho == OSQP_ADAPTIVE_RHO_UPDATE_TIME, this is the fraction of the setup time to use as the rho adaptation interval.

When adaptive_rho == OSQP_ADAPTIVE_RHO_UPDATE_KKT_ERROR, this is the fraction of the previous KKT error to adapt rho at.

Not used for any other adaptive_rho value.

adaptive_rho_tolerance: osqp_float

Tolerance applied when adapting rho.

New rho must be X times larger or smaller than the current one to change it

max_iter: osqp_int

< maximum number of iterations

eps_abs: osqp_float

< absolute solution tolerance

eps_rel: osqp_float

< relative solution tolerance

eps_prim_inf: osqp_float

< primal infeasibility tolerance

eps_dual_inf: osqp_float

< dual infeasibility tolerance

scaled_termination: osqp_int

< boolean; use scaled termination criteria

check_termination: osqp_int

< integer, check termination interval; if 0, checking is disabled

check_dualgap: osqp_int

< Boolean; use duality gap termination criteria

time_limit: osqp_float

< maximum time to solve the problem (seconds)

delta: osqp_float

< regularization parameter for polishing

polish_refine_iter: osqp_int

< number of iterative refinement steps in polishing